Add Client and Server that can connect and communicate with eachother
Add a server that waits for clients to connect and sends a message
to the client when they connect.

Add a client that connects to the server and can receive messages
from the server

Add test code for the client that demonstates how to call the client library
and prints out a message sent from the server.